requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.8.3
pillow==10.1.0
pymupdf==1.23.7
requests==2.31.0
pydantic==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.2
aiohttp==3.9.1

